+++ test/peddler/helpers/reports_2021_06_30_test.rb
@@ -48,17 +48,8 @@ module Peddler
       def test_download_report_document_client_error
         url = "https://tortuga-prod-na.s3-external-1.amazonaws.com/123456"
 
-        begin
-          require "nokogiri"
-          # With nokogiri available, should parse XML error and raise Peddler::Error
-          assert_raises(Peddler::Error) do
-            download_report_document(url)
-          end
-        rescue LoadError
-          # Without nokogiri, XML parsing fails gracefully - raises generic Peddler::Error
-          assert_raises(Peddler::Error) do
-            download_report_document(url)
-          end
+        assert_raises(Peddler::Error) do
+          download_report_document(url)
         end
       end
     end