+++ test/peddler/response_test.rb
@@ -49,21 +49,6 @@ module Peddler
       assert_equal(payload, wrapper.parse)
     end
 
-    def test_decorate_shows_deprecation
-      wrapper = nil
-      assert_output(nil, /Response\.decorate is deprecated/) do
-        wrapper = Response.decorate(response)
-      end
-
-      assert_equal(payload, wrapper.parse)
-    end
-
-    def test_deprecation_removal_reminder
-      if Gem.loaded_specs["peddler"].version.segments.first >= 5
-        flunk("Response.decorate should have been removed in v5.0. Please delete it now.")
-      end
-    end
-
     def test_parsable_with_parser
       wrapper = Response.new(response, parser: custom_parser)